+++ kmfrontend/src/frontend/components/NotfoundPage.tsx
@@ -13,8 +13,8 @@ class NotfoundPage extends Component {
 				<h1>{i18next.t('404')}</h1>
 				<h3>{i18next.t('404_2')}</h3>
 				<div className="you-are-here">{window.location.pathname} &lt;----- {i18next.t('404_3')}</div>
-				<a href="/" className="page404-btn">{i18next.t('404_4')}</a><br/>
-				<img alt="" height="150" src={image404} />
+				<a href="/" className="page404-btn">{i18next.t('404_4')}</a>
+				<div><img alt="" height="150" src={image404} /></div>
 			</div>
 		);
 	}