+++ kmfrontend/src/frontend/components/modals/Tutorial.tsx
@@ -75,7 +75,7 @@ class Tutorial extends Component<unknown, IState> {
 							{i18next.t('MODAL.TUTORIAL.WRENCH_BUTTON')} <i className="fas fa-fw fa-wrench" /> {i18next.t('MODAL.TUTORIAL.WRENCH_BUTTON_DESC')}
 						</p>
 					</div>
-					{i18next.t('MODAL.TUTORIAL.LINES_WITH')} <span className="orange">{i18next.t('MODAL.TUTORIAL.ORANGE')}</span> {i18next.t('MODAL.TUTORIAL.KARAOKE_PLAYED')}<br />
+					{i18next.t('MODAL.TUTORIAL.LINES_WITH')} <span className="orange"><i className="fas fa-fw fa-history" /> {i18next.t('MODAL.TUTORIAL.ORANGE')}</span> {i18next.t('MODAL.TUTORIAL.KARAOKE_PLAYED')}<br />
 					{i18next.t('MODAL.TUTORIAL.LINE_WITH')} <span className="blue">{i18next.t('MODAL.TUTORIAL.BLUE')}</span> {i18next.t('MODAL.TUTORIAL.KARAOKE_PLAYING')}
 				</div>
 			</>;
@@ -89,6 +89,7 @@ class Tutorial extends Component<unknown, IState> {
 						<ul className="ul-l1">
 							<li><i className="fas fa-fw fa-book" /> {i18next.t('MODAL.TUTORIAL.LIBRARY')}<br />
 								{i18next.t('MODAL.TUTORIAL.DOWNLOAD')} <a href="/system/karas/download" target="_blank">{i18next.t('MODAL.TUTORIAL.SYSTEM_PANEL')}</a>.</li>
+							<li><i className="fas fa-fw fa-pencil-alt" /> {i18next.t('MODAL.TUTORIAL.PLAYLIST_ATTRIBUTES')} <strong>{i18next.t('MODAL.TUTORIAL.CURRENT')}</strong> {i18next.t('MODAL.TUTORIAL.OR')} <strong>{i18next.t('MODAL.TUTORIAL.PUBLIC')}</strong>.</li>
 							<li><i className="fas fa-fw fa-play-circle" /> {i18next.t('MODAL.TUTORIAL.THE_PLAYLIST')} <strong>{i18next.t('MODAL.TUTORIAL.CURRENT')}</strong> {i18next.t('MODAL.TUTORIAL.CURRENT_DESC')}</li>
 							<li><i className="fas fa-fw fa-globe" /> {i18next.t('MODAL.TUTORIAL.THE_PLAYLIST')} <strong>{i18next.t('MODAL.TUTORIAL.PUBLIC')}</strong> {i18next.t('MODAL.TUTORIAL.PUBLIC_DESC')}</li>
 							<li><i className="fas fa-fw fa-info-circle" /> {i18next.t('MODAL.TUTORIAL.THE_PLAYLIST')} <strong>{i18next.t('MODAL.TUTORIAL.CURRENT')}</strong> {i18next.t('MODAL.TUTORIAL.AND')} <strong>{i18next.t('MODAL.TUTORIAL.PUBLIC')}</strong> {i18next.t('MODAL.TUTORIAL.CURRENT_PUBLIC_DESC')}</li>
@@ -99,10 +100,8 @@ class Tutorial extends Component<unknown, IState> {
 			break;
 		case 2:
 			slide = <div className="player-presentation">
+				<p>{i18next.t('MODAL.TUTORIAL.PLAYER_BAR')}</p>
 				<ul>
-					<li><i className="fas fa-fw fa-play" />
-						{i18next.t('MODAL.TUTORIAL.PLAYER_BAR')}
-					</li>
 					<li><i className="fas fa-fw fa-play-circle" />
 						{i18next.t('MODAL.TUTORIAL.PLAYER_CURRENT_HINT')}&nbsp;
 						<strong>