+++ CHANGELOG.md
@@ -153,7 +153,7 @@ It can be enabled in the Window menu.
 - Removed option to disallow duplicates of the same series/singer in a playlist.
 - Removed option to display visualization effects during audio-only karaokes. This wasn't very nice-looking and caused issues with some songs. (#934)
 - New icon for macOS 11 Big Sur (#856)
-- Upgraded ffmpeg toolset to 4.3.1 and mpv media player to 0.33 (with libass fixes, #826)
+- Upgraded ffmpeg toolset to 4.3.2 and mpv media player to 0.33.1 (with libass fixes, #826)
 - Upgraded to Typescript 4.0 (#784)
 - Karaoke Mugen 5.0 isn't generating Karaoke Mugen <3.x series files anymore (#738)
 - appPath detection has been reworked (#747)